thêm các logic của phần thêm sửa xóa: - Xóa user thì cũng phải xóa ảnh 				v - Sửa ảnh của user thì cũng phải xóa anh cũ đi		? - Sửa user thì nếu kiểm tra thấy khác tổ thì cũng phải giảm số lượng của tổ ấy đi và tăng ở chỗ tổ mới     v - Sửa criteria thì cx phải kiểm tra xem thấy khác criteriaGroup thif cũng phải giảm số lượng của cũ đi và tăng cái mới 			v - Xóa tổ thì phải check xem có user trong tổ hay không		v - Xóa rank thì phải xem evaluate có không		v - Xóa nhóm tiêu chí thì phải xem có tiêu chí không  	v - Xóa người người thì phải check xem có evaluate của người dùng không thì xóa cả đi.                V
- Xóa tiêu chí phải check xem có evaluate nào k		v

+++ Controllers/CriteriaGroupController.cs
@@ -13,10 +13,12 @@ namespace WebAPIwithMongoDB.Controllers
     public class CriteriaGroupController : ControllerBase
     {
         private readonly ICriteriaGroupRepository _criteriaGroupRepository;
+        private readonly ICriteriaRepository _criteriaRepository;
 
-        public CriteriaGroupController(ICriteriaGroupRepository criteriaGroupRepository)
+        public CriteriaGroupController(ICriteriaGroupRepository criteriaGroupRepository, ICriteriaRepository criteriaRepository)
         {
             _criteriaGroupRepository = criteriaGroupRepository;
+            _criteriaRepository = criteriaRepository;
         }
 
         [HttpGet]
@@ -69,6 +71,12 @@ namespace WebAPIwithMongoDB.Controllers
         {
             if (!await _criteriaGroupRepository.Exists(id))
                 return NotFound(new ApiResponse<string>(404, "Không tìm thấy nhóm tiêu chí", null));
+
+            var criteria = await _criteriaRepository.GetCriteriesByCriteriaGroupId(id);
+            if (criteria.Any())
+            {
+                return BadRequest(new ApiResponse<string>(400, "Xóa không thành công vì còn tiêu chí bên trong", null));
+            }
 
             await _criteriaGroupRepository.DeleteAsync(id);
             return Ok(new ApiResponse<string>(200, "Xóa thành công", null));
